добавлен эндпоинт для проверки введённых данных

+++ server.js
@@ -37,17 +37,6 @@ const generateTokenPair = (data) => {
     }
 }
 
-const getTokenData = (token) => {
-    return new Promise((resolve, reject) => {
-        jwt.verify(token, config.atSecret, (err, decoded) => {
-            if (err) {
-                reject(err)
-            } else {
-                resolve(decoded)
-            }
-        })
-    })
-}
 
 app.use(cors({
     origin: originCors,
@@ -62,13 +51,26 @@ app.post('/reg',async (req, res)=>{
     const newUser = req.body 
     const usersData = fs.readFileSync('usersData.json')
     const parsedUsers = JSON.parse(usersData)
-    parsedUsers.push(newUser) 
-    const newData = JSON.stringify(parsedUsers)
-    fs.writeFileSync('usersData.json', newData)
-    res.status = 200
-    res.send({
-        message: 'Новый пользователь зарегистрирован'
-    })
+    const existedUser = parsedUsers.find((user)=>user.name === newUser.name)
+    console.log(existedUser)
+    if(!existedUser) {
+        parsedUsers.push(newUser) 
+        const newData = JSON.stringify(parsedUsers)
+        console.log('Регистрация нового пользователя')
+        fs.writeFileSync('usersData.json', newData)
+        res.status = 200
+        res.send({
+            name: newUser.name,
+            password: newUser.password,
+            message: 'Новый пользователь зарегистрирован'
+        })
+    } else {
+        console.log('Имя занято')
+        res.status = 200
+        res.send({
+            message: 'Пользователь с таким именем уже существует'
+        })
+    }
 })
 
 app.post('/auth', async (req, res)=>{
@@ -79,29 +81,58 @@ app.post('/auth', async (req, res)=>{
         user.name === authUser.name
     )
     if(existedUser) {
+        console.log(existedUser, authUser, 'пользователь существует')
         if (existedUser.password !== authUser.password) {
+            res.status = 200
             res.send({
-                isUserLoggedIn: false,
                 message: 'Ошибка - неверный пароль'
             })
+            console.log('Неверный пароль')
+        } 
+        if (existedUser.password === authUser.password) {
+            const { accessToken, refreshToken } = generateTokenPair({name: existedUser.name})
+            res.status = 200
+            res.send({
+                message: "Успешно",
+                accessToken,
+                refreshToken
+            })
         }
-        const { accessToken, refreshToken } = generateTokenPair({
-            name: existedUser.name
-        })
-
-        res.status = 200
-        res.send({
-            message: "Успешно",
-            accessToken,
-            refreshToken
-        })
     } else {
+        console.log('Ошибка - пользователя с таким именем не существует')
         res.send({
             message: 'Ошибка - пользователя с таким именем не существует'
         })
     }
 })
 
+app.post('/check',async (req, res)=>{
+    const checkedUser = req.body 
+    const usersData = fs.readFileSync('usersData.json')
+    const parsedUsers = JSON.parse(usersData)
+    const existedUser = parsedUsers.find((user)=>user.name === checkedUser.name)
+    console.log(checkedUser, existedUser)
+    if (existedUser) {
+        console.log('user exists')
+        if(existedUser.password === checkedUser.password) {
+            res.status = 200
+            res.send(existedUser)
+        } else {
+            console.log('atata')
+            res.status = 200
+            res.send({
+                name: existedUser.name,
+                password: null
+            })
+        }
+    } else {
+        res.status = 200
+        res.send({
+            name: null,
+            password: null
+        })
+    }
+})
 
 app.listen('3011', ()=>{
     console.log('server is listening on port 3011')
